+++ mrv2/docs/HISTORY.md
@@ -21,6 +21,8 @@ v0.8.4
 - Added the code and callback to edit the annotations when editing the clips.
 - Added Edit/Audio Gap/Insert and Edit/Audio Gap/Remove to insert or remove
   audio gaps matching the length of the video clip at the current time.
+- Fixed adding a clip without audio to a timeline with audio that ends before
+  the video.  Now a gap is added.
 
 
 v0.8.3